Move LedgerEntrySource schema to new schemas directory
