+++ src/components/Journal/Journal.tsx
@@ -6,9 +6,9 @@ import { useJournal } from '../../hooks/useJournal'
 import { Container } from '../Container'
 import { JournalTable } from '../JournalTable'
 import { JournalTableStringOverrides } from '../JournalTable/JournalTableWithPanel'
-import { LedgerEntrySource } from '../../types/ledger_accounts'
 import { SourceLink } from '../../types/utility/links'
 import { LedgerEntrySourceProvider } from '../../contexts/LedgerEntrySourceContext'
+import { LedgerEntrySourceType } from '../../schemas/ledgerEntrySourceSchemas'
 
 export interface JournalConfig {
   form: {
@@ -24,7 +24,7 @@ export interface JournalProps {
   asWidget?: boolean
   config?: JournalConfig
   stringOverrides?: JournalStringOverrides
-  convertLedgerEntrySourceToSourceLink?: (source: LedgerEntrySource) => SourceLink | undefined
+  convertLedgerEntrySourceToSourceLink?: (source: LedgerEntrySourceType) => SourceLink | undefined
 }
 
 export const JOURNAL_CONFIG: JournalConfig = {
