+++ src/hooks/useProfitAndLoss/useProfitAndLossDetailLines.tsx
@@ -9,7 +9,7 @@ import { get } from '../../api/layer/authenticated_http'
 import { Schema } from 'effect'
 import { ReportingBasis } from '../../types'
 import { toDefinedSearchParameters } from '../../utils/request/toDefinedSearchParameters'
-import { LedgerEntrySourceSchema, PnlDetailLineSchema, PnlDetailLinesDataSchema } from './schemas'
+import { PnlDetailLineSchema, PnlDetailLinesDataSchema } from './schemas'
 
 export const LIST_PNL_DETAIL_LINES_TAG_KEY = '#list-pnl-detail-lines'
 
@@ -34,7 +34,6 @@ type PnlDetailLinesFilterParams = {
 type PnlDetailLinesParams = PnlDetailLinesBaseParams & PnlDetailLinesFilterParams
 
 export type PnlDetailLine = typeof PnlDetailLineSchema.Type
-export type LedgerEntrySourceType = typeof LedgerEntrySourceSchema.Type
 
 export type PnlDetailLinesReturn = typeof PnlDetailLinesDataSchema.Type
