+++ src/index.tsx
@@ -64,3 +64,5 @@ export { BankTransactionsProvider } from './providers/BankTransactionsProvider'
 export { useDataSync } from './hooks/useDataSync'
 
 export { DisplayState, Direction } from './types/bank_transactions'
+
+export type { LedgerEntrySourceType } from './schemas/ledgerEntrySourceSchemas'
