+++ src/views/GeneralLedger/GeneralLedger.tsx
@@ -6,8 +6,8 @@ import { JournalStringOverrides } from '../../components/Journal/Journal'
 import { ProfitAndLoss } from '../../components/ProfitAndLoss'
 import { Toggle } from '../../components/Toggle'
 import { View } from '../../components/View'
-import { LedgerEntrySource } from '../../types/ledger_accounts'
 import { SourceLink } from '../../types/utility/links'
+import { LedgerEntrySourceType } from '../../schemas/ledgerEntrySourceSchemas'
 
 export interface GeneralLedgerStringOverrides {
   title?: string
@@ -27,7 +27,7 @@ export interface GeneralLedgerProps {
   showTitle?: boolean
   stringOverrides?: GeneralLedgerStringOverrides
   chartOfAccountsOptions?: ChartOfAccountsOptions
-  convertLedgerEntrySourceToSourceLink?: (source: LedgerEntrySource) => SourceLink | undefined
+  convertLedgerEntrySourceToSourceLink?: (source: LedgerEntrySourceType) => SourceLink | undefined
 }
 
 export const GeneralLedgerView = ({
